+++ GHEtool/__init__.py
@@ -2,5 +2,4 @@ import pathlib
 
 from GHEtool.main_class import Borefield
 from GHEtool.VariableClasses import CustomGFunction, FluidData, GFunction, GroundData, PipeData, SizingSetup
-
 FOLDER: pathlib.Path = pathlib.Path(__file__).parent  # solve problem with importing GHEtool from sub-folders