+++ GHEtool/gui/gui_data_storage.py
@@ -1,4 +1,2 @@
-
-
 class DataStorage:
     """placeholder class has moved to gui_classes"""