+++ GHEtool/main_class.py
@@ -4,13 +4,18 @@ This file contains all the code for the borefield calculations.
 from __future__ import annotations
 
 from math import pi
+from typing import Tuple
 
 import matplotlib.pyplot as plt
 import numpy as np
 import pygfunction as gt
-from GHEtool.VariableClasses import CustomGFunction, FluidData, GFunction, GroundData, PipeData, SizingSetup, load_custom_gfunction
-from GHEtool.VariableClasses.BaseClass import BaseClass
+
 from scipy.signal import convolve
+
+from GHEtool.VariableClasses import FluidData, PipeData, Borehole, GroundConstantTemperature
+from GHEtool.VariableClasses import CustomGFunction, load_custom_gfunction, GFunction, SizingSetup
+from GHEtool.VariableClasses.BaseClass import BaseClass
+from GHEtool.VariableClasses.GroundData._GroundData import _GroundData
 
 
 class Borefield(BaseClass):
@@ -26,11 +31,11 @@ class Borefield(BaseClass):
 
     HOURLY_LOAD_ARRAY: np.ndarray = np.arange(0, 8761, UPM).astype(np.uint32)
 
-    __slots__ = 'baseload_heating', 'baseload_cooling', 'H', 'H_init', 'Rb', 'ty', 'tm', \
+    __slots__ = 'baseload_heating', 'baseload_cooling', 'H', 'H_init', 'ty', 'tm', 'borehole',\
                 'hourly_heating_load', 'hourly_cooling_load', 'number_of_boreholes', '_borefield', 'cost_investment', \
                 'length_peak', 'th', 'Tf_max', 'Tf_min', 'limiting_quadrant', 'monthly_load', 'monthly_load_heating', \
                 'monthly_load_cooling', 'peak_heating', 'imbalance', 'qa', 'Tf', 'qm', 'qh', 'qpm', 'tcm', 'tpm', \
-                'peak_cooling', 'simulation_period', 'ground_data', 'pipe_data', 'fluid_data',\
+                'peak_cooling', 'simulation_period', 'ground_data',\
                 'results_peak_heating', 'time_L4', 'options_pygfunction',\
                 'results_peak_cooling', 'results_month_cooling', 'results_month_heating', 'Tb', 'THRESHOLD_WARNING_SHALLOW_FIELD', \
                 'gui', 'time_L3_last_year', 'peak_heating_external', 'peak_cooling_external', \
@@ -174,9 +179,8 @@ class Borefield(BaseClass):
 
         # initiate ground parameters
         self.H = 0.  # borehole depth m
-        self.Rb = 0.  # effective borehole thermal resistance mK/W
         self.number_of_boreholes = 0  # number of total boreholes #
-        self.ground_data: GroundData = GroundData()
+        self.ground_data: _GroundData = GroundConstantTemperature()
         self.D: float = 0.  # buried depth of the borehole [m]
         self.r_b: float = 0.  # borehole radius [m]
 
@@ -184,10 +188,9 @@ class Borefield(BaseClass):
         self.Tf: float = 0.  # temperature of the fluid
         self.Tf_max: float = 16.  # maximum temperature of the fluid
         self.Tf_min: float = 0.  # minimum temperature of the fluid
-        self.fluid_data: FluidData = FluidData()
 
-        # initiate borehole parameters
-        self.pipe_data: PipeData = PipeData()
+        # initiale borehole
+        self.borehole = Borehole()
 
         # initiate different sizing
         self._sizing_setup: SizingSetup = SizingSetup()
@@ -500,7 +503,35 @@ class Borefield(BaseClass):
             # 16 bit is not enough, go to 32
             self.time_L4 = 3600 * np.arange(1, 8760 * self.simulation_period + 1, dtype=np.float32)
 
-    def set_ground_parameters(self, data: GroundData) -> None:
+    @property
+    def Rb(self) -> float:
+        """
+        This function returns the equivalent borehole thermal resistance.
+
+        Returns
+        -------
+        Rb : float
+            Equivalent borehole thermal resistance [mK/W]
+        """
+        return self.borehole._Rb
+
+    @Rb.setter
+    def Rb(self, Rb: float) -> None:
+        """
+        This function sets the equivalent borehole thermal resistance.
+
+        Parameters
+        ----------
+        Rb : float
+            Equivalent borehole thermal resistance [mk/W]
+
+        Returns
+        -------
+        None
+        """
+        self.set_Rb(Rb)
+
+    def set_ground_parameters(self, data: _GroundData) -> None:
         """
         This function sets the relevant ground parameters.
 
@@ -513,7 +544,6 @@ class Borefield(BaseClass):
         -------
         None
         """
-        self.Rb: float = data.Rb
 
         # Ground properties
         self.ground_data = data
@@ -534,10 +564,7 @@ class Borefield(BaseClass):
         -------
         None
         """
-        self.fluid_data = data
-
-        if self.pipe_data.check_values():
-            self.calculate_fluid_thermal_resistance()
+        self.borehole.fluid_data = data
 
     def set_pipe_parameters(self, data: PipeData) -> None:
         """
@@ -552,12 +579,22 @@ class Borefield(BaseClass):
         -------
         None
         """
-        self.pipe_data = data
+        self.borehole.pipe_data = data
 
-        # calculate the different resistances
-        if self.fluid_data.check_values():
-            self.calculate_fluid_thermal_resistance()
-        self.pipe_data.calculate_pipe_thermal_resistance()
+    def set_Rb(self, Rb: float) -> None:
+        """
+        This function sets the equivalent borehole thermal resistance.
+
+        Parameters
+        ----------
+        Rb : float
+            Equivalent borehole thermal resistance (mK/W)
+
+        Returns
+        -------
+        None
+        """
+        self.borehole._Rb = Rb
 
     def set_max_ground_temperature(self, temp: float) -> None:
         """
@@ -589,25 +626,6 @@ class Borefield(BaseClass):
         """
         self.Tf_min: float = temp
 
-    def calculate_fluid_thermal_resistance(self) -> None:
-        """
-        This function calculates and sets the fluid thermal resistance R_f.
-
-        Returns
-        -------
-        None
-        """
-        self.fluid_data.h_f: float =\
-            gt.pipes.convective_heat_transfer_coefficient_circular_pipe(self.fluid_data.mfr /
-                                                                        self.pipe_data.number_of_pipes,
-                                                                        self.pipe_data.r_in,
-                                                                        self.fluid_data.mu,
-                                                                        self.fluid_data.rho,
-                                                                        self.fluid_data.k_f,
-                                                                        self.fluid_data.Cp,
-                                                                        self.pipe_data.epsilon)
-        self.fluid_data.R_f: float = 1. / (self.fluid_data.h_f * 2 * pi * self.pipe_data.r_in)
-
     @property
     def _Rb(self) -> float:
         """
@@ -626,10 +644,21 @@ class Borefield(BaseClass):
         # calculate Rb*
         return self.calculate_Rb()
 
+    def calculate_Rb(self) -> float:
+        """
+        This function calculates the equivalent borehole thermal resistance by calling the calculate_Rb
+        function of the borehole class.
+
+        Returns
+        -------
+        Rb : float
+            Equivalent borehole thermal resistance [mk/W]
+        """
+        return self.borehole.calculate_Rb(self.H, self.D, self.r_b, self.ground_data.k_s)
+
     def _Tg(self, H: float = None) -> float:
         """
-        This function gives back the ground temperature
-        When use_constant_Tg is False, the thermal heat flux is taken into account.
+        This function gives back the ground temperature.
 
         Parameters
         ----------
@@ -644,35 +673,7 @@ class Borefield(BaseClass):
         if H is None:
             H = self.H
 
-        return self.ground_data.calculate_Tg(H, use_constant_Tg=self._sizing_setup.use_constant_Tg)
-
-    def calculate_Rb(self) -> float:
-        """
-        This function returns the calculated equivalent borehole thermal resistance Rb* value.
-
-        Returns
-        -------
-        Rb* : float
-            Equivalent borehole thermal resistance [mK/W]
-
-        Raises
-        ------
-        ValueError
-            ValueError when no pipe or fluid data is available.
-        """
-        # check if all data is available
-        if not self.pipe_data.check_values() or not self.fluid_data.check_values():
-            print("Please make sure you set al the pipe and fluid data.")
-            raise ValueError
-
-        # initiate temporary borefield
-        borehole = gt.boreholes.Borehole(self.H, self.D, self.r_b, 0, 0)
-        # initiate pipe
-        pipe = gt.pipes.MultipleUTube(self.pipe_data.pos, self.pipe_data.r_in, self.pipe_data.r_out,
-                                      borehole, self.ground_data.k_s, self.pipe_data.k_g,
-                                      self.pipe_data.R_p + self.fluid_data.R_f, self.pipe_data.number_of_pipes, J=2)
-
-        return pipe.effective_borehole_thermal_resistance(self.fluid_data.mfr, self.fluid_data.Cp)
+        return self.ground_data.calculate_Tg(H)
 
     @property
     def _Ahmadfard(self) -> float:
@@ -761,7 +762,7 @@ class Borefield(BaseClass):
             self.H = L / self.number_of_boreholes
         return self.H
 
-    def sizing_setup(self, H_init: float = 100, use_constant_Rb: bool = None, use_constant_Tg: bool = None, quadrant_sizing: int = 0,
+    def sizing_setup(self, H_init: float = 100, use_constant_Rb: bool = None, quadrant_sizing: int = 0,
                      L2_sizing: bool = None, L3_sizing: bool = None, L4_sizing: bool = None, sizing_setup: SizingSetup = None) -> None:
         """
         This function sets the options for the sizing function.
@@ -778,8 +779,6 @@ class Borefield(BaseClass):
             Initial depth of the borefield to start the iteration (m)
         use_constant_Rb : bool
             True if a constant borehole equivalent resistance (Rb*) value should be used
-        use_constant_Tg : bool
-            True if a constant Tg value should be used (the geothermal flux is neglected)
         quadrant_sizing : int
             Differs from 0 when a sizing in a certain quadrant is desired.
             Quadrants are developed by (Peere et al., 2021) [#PeereBS]_, [#PeereThesis]_
@@ -810,14 +809,13 @@ class Borefield(BaseClass):
             return
 
         self._sizing_setup = SizingSetup(use_constant_Rb=use_constant_Rb,
-                                         use_constant_Tg=use_constant_Tg,
                                          quadrant_sizing=quadrant_sizing,
                                          L2_sizing=L2_sizing,
                                          L3_sizing=L3_sizing,
                                          L4_sizing=L4_sizing)
 
-    def size(self, H_init: float = 100, use_constant_Rb: bool = None, use_constant_Tg: bool = None,
-             L2_sizing: bool = None, L3_sizing: bool = None, L4_sizing: bool = None, quadrant_sizing: int = None) -> float:
+    def size(self, H_init: float = 100, use_constant_Rb: bool = None, L2_sizing: bool = None,
+             L3_sizing: bool = None, L4_sizing: bool = None, quadrant_sizing: int = None) -> float:
         """
         This function sets the options for the sizing function.
 
@@ -837,8 +835,6 @@ class Borefield(BaseClass):
             Initial depth of the borefield to start the iteration (m)
         use_constant_Rb : bool
             True if a constant borehole equivalent resistance (Rb*) value should be used
-        use_constant_Tg : bool
-            True if a constant Tg value should be used (the geothermal flux is neglected)
         quadrant_sizing : int
             Differs from 0 when a sizing in a certain quadrant is desired.
             Quadrants are developed by (Peere et al., 2021) [#PeereBS]_, [#PeereThesis]_
@@ -867,7 +863,7 @@ class Borefield(BaseClass):
         self._sizing_setup.make_backup()
 
         # run the sizing setup
-        self._sizing_setup.update_variables(use_constant_Rb=use_constant_Rb, use_constant_Tg=use_constant_Tg,
+        self._sizing_setup.update_variables(use_constant_Rb=use_constant_Rb,
                                             L2_sizing=L2_sizing, L3_sizing=L3_sizing, L4_sizing=L4_sizing,
                                             quadrant_sizing=quadrant_sizing)
 
@@ -1590,16 +1586,16 @@ class Borefield(BaseClass):
         ax.set_ylabel(r'Temperature ($^\circ C$)', color="white" if self.gui else "black")
 
         # plot Temperatures
-        ax.step(time_array, self.Tb, 'k-', where="pre", lw=1.5, label="Tb")
+        ax.step(time_array, self.Tb, 'k-', where="post", lw=1.5, label="Tb")
 
         if plot_hourly:
-            ax.step(time_array, self.results_peak_cooling, 'b-', where="pre", lw=1, label='Tf')
+            ax.step(time_array, self.results_peak_cooling, 'b-', where="post", lw=1, label='Tf')
         else:
-            ax.step(time_array, self.results_peak_cooling, 'b-', where="pre", lw=1.5, label='Tf peak cooling')
-            ax.step(time_array, self.results_peak_heating, 'r-', where="pre", lw=1.5, label='Tf peak heating')
+            ax.step(time_array, self.results_peak_cooling, 'b-', where="post", lw=1.5, label='Tf peak cooling')
+            ax.step(time_array, self.results_peak_heating, 'r-', where="post", lw=1.5, label='Tf peak heating')
 
-            ax.step(time_array, self.results_month_cooling, color='b', linestyle="dashed", where="pre", lw=1.5, label='Tf base cooling')
-            ax.step(time_array, self.results_month_heating, color='r', linestyle="dashed", where="pre", lw=1.5, label='Tf base heating')
+            ax.step(time_array, self.results_month_cooling, color='b', linestyle="dashed", where="post", lw=1.5, label='Tf base cooling')
+            ax.step(time_array, self.results_month_heating, color='r', linestyle="dashed", where="post", lw=1.5, label='Tf base heating')
 
         # define temperature bounds
         ax.hlines(self.Tf_min, 0, self.simulation_period, colors='r', linestyles='dashed', label='', lw=1)
@@ -1759,7 +1755,7 @@ class Borefield(BaseClass):
             1D array with the g-values for all the requested time_value(s)
         """
         # when using a variable ground temperature, sometimes no solution can be found
-        if not self._sizing_setup.use_constant_Tg and H > Borefield.THRESHOLD_DEPTH_ERROR:
+        if not isinstance(self.ground_data, GroundConstantTemperature) and H > Borefield.THRESHOLD_DEPTH_ERROR:
             raise ValueError("Due to the use of a variable ground temperature, no solution can be found."
                              "To see the temperature profile, one can plot it using the depth of ",
                              str(Borefield.THRESHOLD_DEPTH_ERROR), "m.")
@@ -2194,6 +2190,51 @@ class Borefield(BaseClass):
             return 3
         return 2
 
+    def plot_load_duration(self, legend: bool = False) -> Tuple[plt.Figure, plt.Axes]:
+        """
+        This function makes a load-duration curve from the hourly values.
+
+        Parameters
+        ----------
+        legend : bool
+            True if the figure should have a legend
+
+        Returns
+        ----------
+        Tuple
+            plt.Figure, plt.Axes
+        """
+        # check if there are hourly values
+        self._check_hourly_load()
+
+        # sort heating and cooling load
+        heating = self.hourly_heating_load.copy()
+        heating[::-1].sort()
+
+        cooling = self.hourly_cooling_load.copy()
+        cooling.sort()
+        cooling = cooling * (-1)
+        # create new figure and axes if it not already exits otherwise clear it.
+        fig = plt.figure()
+        ax = fig.add_subplot(111)
+        # add sorted loads to plot
+        ax.step(np.arange(0, 8760, 1), heating, 'r-', label="Heating")
+        ax.step(np.arange(0, 8760, 1), cooling, 'b-', label="Cooling")
+        # create 0 line
+        ax.hlines(0, 0, 8759, color="black")
+        # add labels
+        ax.set_xlabel("Time [hours]")
+        ax.set_ylabel("Power [kW]")
+        # set x limits to 8760
+        ax.set_xlim(0, 8760)
+        # plot legend if wanted
+        if legend:
+            ax.legend()
+        # show plt if not in gui
+        if not self.gui:
+            plt.show()
+        return fig, ax
+
     def draw_borehole_internal(self) -> None:
         """
         This function draws the internal structure of a borehole.
@@ -2205,7 +2246,7 @@ class Borefield(BaseClass):
         """
 
         # calculate the pipe positions
-        pos = self.pipe_data._axis_symmetrical_pipe
+        pos = self.borehole.pipe_data._axis_symmetrical_pipe
 
         # set figure
         figure, axes = plt.subplots()
@@ -2215,11 +2256,11 @@ class Borefield(BaseClass):
         circles_inner = []
 
         # color inner circles and outer circles
-        for i in range(self.pipe_data.number_of_pipes):
-            circles_outer.append(plt.Circle(pos[i], self.pipe_data.r_out, color="black"))
-            circles_inner.append(plt.Circle(pos[i], self.pipe_data.r_in, color="red"))
-            circles_outer.append(plt.Circle(pos[i + self.pipe_data.number_of_pipes], self.pipe_data.r_out, color="black"))
-            circles_inner.append(plt.Circle(pos[i + self.pipe_data.number_of_pipes], self.pipe_data.r_in, color="blue"))
+        for i in range(self.borehole.pipe_data.number_of_pipes):
+            circles_outer.append(plt.Circle(pos[i], self.borehole.pipe_data.r_out, color="black"))
+            circles_inner.append(plt.Circle(pos[i], self.borehole.pipe_data.r_in, color="red"))
+            circles_outer.append(plt.Circle(pos[i + self.borehole.pipe_data.number_of_pipes], self.borehole.pipe_data.r_out, color="black"))
+            circles_inner.append(plt.Circle(pos[i + self.borehole.pipe_data.number_of_pipes], self.borehole.pipe_data.r_in, color="blue"))
 
         # set visual settings for figure
         axes.set_aspect('equal')
@@ -2246,47 +2287,3 @@ class Borefield(BaseClass):
 
         # show plot
         plt.show()
-
-    def plot_load_duration(self, legend: bool = False) -> tuple[plt.Figure, plt.Axes]:
-        """
-        This function makes a load-duration curve from the hourly values.
-
-        Parameters
-        ----------
-        legend : bool
-            True if the figure should have a legend
-
-        Returns
-        ----------
-        Tuple
-            plt.Figure, plt.Axes
-        """
-        # check if there are hourly values
-        self._check_hourly_load()
-        # sort heating and cooling load
-        heating = self.hourly_heating_load.copy()
-        heating[::-1].sort()
-
-        cooling = self.hourly_cooling_load.copy()
-        cooling.sort()
-        cooling = cooling * (-1)
-        # create new figure and axes if it not already exits otherwise clear it.
-        fig = plt.figure()
-        ax = fig.add_subplot(111)
-        # add sorted loads to plot
-        ax.step(np.arange(0, 8760, 1), heating, 'r-', label="Heating")
-        ax.step(np.arange(0, 8760, 1), cooling, 'b-', label="Cooling")
-        # create 0 line
-        ax.hlines(0, 0, 8759, color="black")
-        # add labels
-        ax.set_xlabel("Time [hours]", color="white" if self.gui else "black")
-        ax.set_ylabel("Power [kW]", color="white" if self.gui else "black")
-        # set x limits to 8760
-        ax.set_xlim(0, 8760)
-        # plot legend if wanted
-        if legend:
-            ax.legend()
-        # show plt if not in gui
-        if not self.gui:
-            plt.show()
-        return fig, ax