+++ CHANGELOG.md
@@ -6,9 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [2.1.2] - [expected] apr 2023
 
 ### Added
-- Coaxial pipes
-- Variable temperature sizing (at least in the code version)
-- Reimplemented size by length and width
+- Different classes for the ground data: one for constant ground temperature and one for the ground temperature with flux (issue #45).
+
+### Changed
+- In figure plotting, the interval[x[i], x[i+1]) now has the value y[i] (instead of y[i-1]).
+- Scroll behaviour on the result page (issue #99).
+- Changed icon of GHEtool.
 
 ### Fixed
 - Wrong heating load in april in GUI (issue #94).